Do not use well known labels for overrides
Do not use well known labels to configure node overrides like
 * openstack-control-plane=enabled
 * openstack-compute-node=enabled
 * openstack-gateway=enabled

We should never target overrides on this labels as this
makes impossible to add additional overrides on nodes with specific class.
For example if we use openstack-compute-node=enabled than we don't have
ability to setup additional overrides for compute nodes.

This patch does not limit using well known labels in osdpl
but this will be done in future.

Related-Prod: PRODX-48199
Change-Id: I7ba91a8e83f1dfe371c13cb2482ed72e432c24d6