[ovn][migration] Add check for VLAN networks
It is known that OVN has multiple issues related to
vlan networks [1], [2], [3]. The same problems are
related to flat networks.
So generally migrating is not recommended for environments
with problematic networks. The critical error will be shown
in case workloads are present in these networks and
routing is configured for such network. In other cases
warning will be raised.

[1] https://bugs.launchpad.net/neutron/+bug/2028846
[2] https://bugs.launchpad.net/octavia/+bug/2095807
[3] https://bugs.launchpad.net/neutron/+bug/1875852

Related-prod: https://mirantis.jira.com/browse/PRODX-49612
Change-Id: Ied5b29242a08952f0822857b1ef4aec7051731f1