Turn jsonified string to proper Python for Horizon
currently casting config.raw to JSON in fact does not work nicely
when values contain some values, like booleans or null/None,
as JSON and Python differ very much in this part of syntax.

To fix this, we add additional json.loads of value,
and wrap the string in tripple quotes + space
to not really bother with escaping quotes in JSON string.

Related-Issue: PRODX-52366
Change-Id: I585b0b4de41af44cc6a2bc20de83e9e12c44b198